♻️ CadastralReferenceField extracted to its own file
